Add files via upload
This project is a backend application for a voting system, enabling users to securely cast votes for candidates. It includes key functionalities such as user authentication to ensure secure access, candidate management for adding or updating candidates, and voting mechanisms to track user votes. Built using Node.js and Express.js, the application utilizes MongoDB as the database for efficient data storage and retrieval. JSON Web Tokens (JWT) are employed for robust authentication, ensuring a secure user experience. This project aims to provide a reliable, scalable, and easy-to-use backend for voting applications, emphasizing security, efficiency, and seamless functionality.